ConfigLoader refactoring
