+++ pieapp/containers/content-table/plugin/api/structs.py
@@ -2,7 +2,7 @@
 ContentTable structs
 """
 
-from piekit.system.config import Config
+from piekit.system import Config
 
 
 DEFAULT_CONTENT_TABLE_COLUMNS: tuple = (
