--- piekit/system/config/__init__.py
@@ -1,1 +1,0 @@
-from piekit.system.config.loader import Config
